ft(support-close-partner): added the support of dynamic keys to interact with guhemba
